+++ Server/src/index.ts
@@ -3,10 +3,12 @@ import express from "express";
 import authRouter from "./routes/auth-route";
 import userRouter from "./routes/user-route";
 import eventRouter from "./routes/event-route";
+import stats from "./routes/stats-route";
+
 import { verifyToken } from "./middlewares/auth-middleware";
 import { error } from "./middlewares/error-middleware";
 import { notFound } from "./middlewares/not-found-middleware";
-import wallet from "./routes/wallet-route";
+
 import cookies from "cookie-parser";
 
 const PORT = process.env.PORT || 8069;
@@ -20,7 +22,7 @@ app.use(cookies());
 app.use("/api/v1/auth", authRouter);
 app.use("/api/v1/users", userRouter);
 app.use("/api/v1/events", verifyToken, eventRouter);
-app.use("/api/v1/wallet", wallet);
+app.use("/api/v1/stats", stats);
 
 app.use(notFound);
 app.use(error);
