+++ Server/src/routes/stats-route.ts
@@ -1,0 +1,8 @@
+import express from "express";
+import { getStatistic } from "../controllers/statistic-controller";
+
+const router = express.Router();
+
+router.route("/").get(getStatistic);
+
+export default router;
